+++ templates/contacts/index.php
@@ -272,7 +272,7 @@ $ownFields = $ownContact !== null ? [
                 <a class="ghost-button" href="<?= e(url('/rundmail?' . http_build_query(array_merge($filters, ['from' => 'filter'])))) ?>"><?= icon('mail') ?><span>Rundmail an diese Liste</span></a>
             <?php endif; ?>
             <?php if ($canManage): ?>
-                <a class="ghost-button" href="<?= e(url('/namensliste' . (($filters['category_id'] ?? '') !== '' ? '?category_id=' . rawurlencode((string) $filters['category_id']) : ''))) ?>"><?= icon('contacts') ?><span>Namensliste</span></a>
+                <a class="ghost-button" href="<?= e(url('/vollstaendigkeit' . (($filters['category_id'] ?? '') !== '' ? '?category_id=' . rawurlencode((string) $filters['category_id']) : ''))) ?>"><?= icon('check') ?><span>Vollständigkeit</span></a>
             <?php endif; ?>
             <?php if (can('contacts.export')): ?>
                 <a class="ghost-button" href="<?= e(url('/contacts/export?' . http_build_query($filters))) ?>"><?= icon('upload') ?><span>CSV exportieren</span></a>